add additionalItems to decode, encode, validation, and fuzz
add tests of additionalItems to SchemaFuzzSpec

NOTE: the fuzzer is incomplete in that it does not take into account minItems
and maxItems constraints. But these can be quite hairy to deal with depending on
the presence and length of `items` and presence of `additionalItems`. There is a
failing test left in for this.