reimplement tuple schema fuzzer to account for minItems, maxItems constraints
add many fuzz test cases for tuple schemas

Note there is the possibility of impossible state when minItems > maxItems, etc.
For these cases the fuzzer just emits an empty value on the assumption they
should never actually happen. But it kind of sucks to let impossible states in
like this.